+++ src/lib/components/index.ts
@@ -3,6 +3,7 @@ export { default as AuthFlow } from "./AuthFlow.svelte";
 export { default as AuthPill } from "./AuthPill.svelte";
 export { default as OtpInput } from "./OtpInput.svelte";
 export { default as PasskeyPrompt } from "./PasskeyPrompt.svelte";
+export { default as SessionRenew } from "./SessionRenew.svelte";
 export type { AuthMessages } from "../i18n/types.js";
 export {
   resolveMessages,